Update CMakefile, rearrange directory.
