--- HMM/tests/non_polyA_train.fa
@@ -1,1 +1,0 @@
-non_polyA_train_real.fa
