--- HMM/tests/non_polyA_train_dummy.fa
@@ -1,10 +1,0 @@
->A
-AA
->C
-C
->G
-G
->T
-T
->T
-T
