--- HMM/tests/polyA_train.fa
@@ -1,1 +1,0 @@
-polyA_train_real.fa
