--- HMM/tests/polyA_train_dummy.fa
@@ -1,8 +1,0 @@
->A
-AAAAAAAAAA
->C
-C
->G
-G
->T
-TT
