+++ aux/C/extractSoftclipped.c
