+++ aux/R/trimpolyA.R
