+++ aux/python/isoseq_draft.primer_info.csv
