+++ aux/python/pacb_extract_insert.py
