+++ src/fasta.hpp
@@ -34,89 +34,67 @@
 // SUCH DAMAGE.
 
 // Author: Bo Han
+#ifndef fasta_h
+#define fasta_h
+
 #include <string>
-#include "gmock/gmock.h"
+#include <iostream>
+#include <cassert>
+#include <fstream>
+#include <memory>
+#include "format.hpp"
 #include "sequence.hpp"
+#include "type_policy.h"
 
-using namespace std;
-namespace {
-class SequenceTest : public ::testing::Test {
-protected:
-    SequenceTest()
-        : seq("GGATCGATCcatcga")
-    {
-    }
-
-public:
-    Sequence<> seq;
+template <class T = caseInsensitiveString>
+struct Fasta : public Sequence<T> {
+    using seq_type = Sequence<T>;
+    std::string name_; // sequnece can take different representation such as char*, std::string or twoBits; name has to be string
 };
 
-TEST_F(SequenceTest, Constructor1)
-{
-    Sequence<> seq2{ "GGATCGATCCATCGA" };
-    fprintf(stderr, "%s\n", seq.seq_.c_str());
-    fprintf(stderr, "%s\n", seq2.seq_.c_str());
-    EXPECT_TRUE(seq == seq2);
-}
+/* reading policy */
+template <class T>
+struct read_policy<Fasta<T> > {
+    using string_type = T;
+    using fasta_type = Fasta<T>;
+    // reading policy for fasta in the ifstream
+    static fasta_type read(std::istream* ins)
+    {
+        fasta_type fa{};
+        if (ins->peek() != '>' || ins->eof()) {
+            return fa; // returning an empty Fa meant EOF or ill-formated file
+        }
+        ins->get(); // consume '>'
+        std::getline(*ins, fa.name_); // read name, which is always std::string
+        constexpr size_t bufsize = 20480;
+        char buffer[bufsize];
+        while (ins->peek() != '>' && ins->good()) {
+            do {
+                ins->clear();
+                ins->getline(buffer, bufsize, '\n');
+                fa.seq_ += buffer;
+            } while ((bufsize == ins->gcount() + 1) && (ins->rdstate() == std::ios::failbit));
+        }
+        return fa;
+    }
+};
 
-TEST_F(SequenceTest, Reverse)
-{
-    Sequence<> seq2{ "agctacCTAGCTAGG" };
-    seq.reverse();
-    EXPECT_TRUE(seq == seq2);
-}
+template <class T>
+struct FastaSupported {
+    const static bool value = false;
+};
 
-TEST_F(SequenceTest, ReverseCopy)
-{
-    Sequence<> seq2{ "agctacCTAGCTAGG" };
-    auto seq3 = seq.reverse_copy();
-    EXPECT_TRUE(seq3 == seq2);
-}
+template <>
+struct FastaSupported<std::string> {
+    const static bool value = true;
+}; // currently string is supported
 
-TEST_F(SequenceTest, Complement1)
-{
-    Sequence<> seq2{ "CCTAGCTAGGTAGCT" };
-    seq.complement();
-    EXPECT_TRUE(seq == seq2);
-}
+template <>
+struct FastaSupported<caseInsensitiveString> {
+    const static bool value = true;
+}; // currently caseInsensitiveString is supported
 
-TEST_F(SequenceTest, Complement2)
-{
-    Sequence<> seq2{ "CCTAGCTAGGTAGC" };
-    Sequence<> seq3{ "GGATCGATCcatcg" };
-    seq2.complement();
-    EXPECT_TRUE(seq3 == seq2);
-}
+template <class T = caseInsensitiveString, class = typename std::enable_if<FastaSupported<T>::value>::type>
+using FastaReader = FormatReader<Fasta<T> >;
 
-TEST_F(SequenceTest, ComplementCopy)
-{
-    auto seq2 = seq.complement_copy();
-    Sequence<> seq3{ "CCTAGCTAGGTAGCT" };
-    EXPECT_TRUE(seq3 == seq2);
-}
-
-TEST_F(SequenceTest, ReverseComplement)
-{
-    Sequence<> seq2{ "TCGATGGATCGATCC" };
-    seq.reverse_complement();
-    EXPECT_TRUE(seq == seq2);
-}
-
-TEST_F(SequenceTest, ReverseComplementCopy)
-{
-    Sequence<> seq2{ "TCGATGGATCGATCC" };
-    auto seq3 = seq.reverse_complement_copy();
-    EXPECT_TRUE(seq3 == seq2);
-}
-
-TEST_F(SequenceTest, Size)
-{
-    EXPECT_TRUE(seq.size() == seq.seq_.size());
-}
-}
-
-int main(int argc, char** argv)
-{
-    testing::InitGoogleMock(&argc, argv);
-    return RUN_ALL_TESTS();
-}
+#endif
