+++ src/fastq.hpp
@@ -34,8 +34,8 @@
 // SUCH DAMAGE.
 
 // Author: Bo Han
-#ifndef fasta_h
-#define fasta_h
+#ifndef fastq_h
+#define fastq_h
 
 #include <string>
 #include <iostream>
@@ -47,54 +47,59 @@
 #include "type_policy.h"
 
 template <class T = caseInsensitiveString>
-struct Fasta : public Sequence<T> {
+struct Fastq : public Sequence<T> {
     using seq_type = Sequence<T>;
-    std::string name_; // sequnece can take different representation such as char*, std::string or twoBits; name has to be string
+    std::string name_;
+    std::string quality_;
 };
 
 /* reading policy */
 template <class T>
-struct read_policy<Fasta<T> > {
+struct read_policy<Fastq<T> > {
+    using traits_type = typename T::traits_type;
     using string_type = T;
-    using fasta_type = Fasta<T>;
+    using fastq_type  = Fastq<T>;
     // reading policy for fasta in the ifstream
-    static fasta_type read(std::istream* ins)
+    static fastq_type read(std::istream* ins)
     {
-        fasta_type fa{};
-        if (ins->peek() != '>' || ins->eof()) {
-            return fa; // returning an empty Fa meant EOF or ill-formated file
+        fastq_type fq{};
+        if (ins->peek() != '@' || ins->eof()) {
+            return fq; // returning an empty Fa meant EOF or ill-formated file
         }
-        ins->get(); // consume '>'
-        std::getline(*ins, fa.name_); // read name, which is always std::string
-        constexpr size_t bufsize = 20480;
-        char buffer[bufsize];
-        while (ins->peek() != '>' && ins->good()) {
-            do {
-                ins->clear();
-                ins->getline(buffer, bufsize, '\n');
-                fa.seq_ += buffer;
-            } while ((bufsize == ins->gcount() + 1) && (ins->rdstate() == std::ios::failbit));
+        ins->get(); // consume '@'
+        std::getline(*ins, fq.name_); // read name, which is always std::string
+        char c;
+        while(true) {
+            ins->get(c);
+            if(c == '\n') break;
+            fq.seq_ += c;
         }
-        return fa;
+        ins->ignore(std::numeric_limits<int>::max(), '\n');
+        std::getline(*ins, fq.quality_);
+        if(fq.seq_.size() != fq.quality_.size()) {
+            fprintf(stderr, "[warning] the length of sequence and quality does not match for %s\n",
+                    fq.name_.c_str());
+        }
+        return fq;
     }
 };
 
 template <class T>
-struct FastaSupported {
+struct FastqSupported {
     const static bool value = false;
 };
 
 template <>
-struct FastaSupported<std::string> {
+struct FastqSupported<std::string> {
     const static bool value = true;
 }; // currently string is supported
 
 template <>
-struct FastaSupported<caseInsensitiveString> {
+struct FastqSupported<caseInsensitiveString> {
     const static bool value = true;
 }; // currently caseInsensitiveString is supported
 
-template <class T = caseInsensitiveString, class = typename std::enable_if<FastaSupported<T>::value>::type>
-using FastaReader = FormatReader<Fasta<T> >;
+template <class T = caseInsensitiveString, class = typename std::enable_if<FastqSupported<T>::value>::type>
+using FastqReader = FormatReader<Fastq<T> >;
 
 #endif
