+++ src/hmm_utilities.h
