+++ src/kernel_color.h
