+++ src/polyA_hmm_model.hpp
@@ -68,74 +68,22 @@ public:
     };
     // methods
 public:
-    PolyAHmmMode()
-        : _base(nStates, nSymbol)
-    {
-    }
-    virtual ~PolyAHmmMode() {}
-    PolyAHmmMode(const PolyAHmmMode& other)
-        : _base(other)
-        , forw_(other.forw_)
-        , back_(other.back_)
-        , post_(other.post_)
-        , path_(other.path_)
-    {
-    }
+    PolyAHmmMode();
 
-    PolyAHmmMode(PolyAHmmMode&& other)
-        : _base(std::move(other))
-        , forw_(std::move(other.forw_))
-        , back_(std::move(other.back_))
-        , post_(std::move(other.post_))
-        , path_(std::move(other.path_))
-    {
-    }
+    virtual ~PolyAHmmMode() {}
+
+    PolyAHmmMode(const PolyAHmmMode& other);
+
+    PolyAHmmMode(PolyAHmmMode&& other);
 
     PolyAHmmMode& operator=(const PolyAHmmMode&) = delete;
 
-    PolyAHmmMode& operator=(PolyAHmmMode&& other)
-    {
-        if (this != &other) {
-            _base::operator=(std::move(other));
-            forw_ = std::move(other.forw_);
-            back_ = std::move(other.back_);
-            post_ = std::move(other.post_);
-            path_ = std::move(other.path_);
-        }
-        return *this;
-    }
+    PolyAHmmMode& operator=(PolyAHmmMode&& other);
 
-    virtual bool read(const std::string& filename)
-    {
-        bool ret = _base::read(filename);
-        // any chance of overwriting
-        //        init_[States::POLYA] = 0.8;
-        //        init_[States::NONPOLYA] = 1 - init_[States::POLYA];
-        //        tran_(States::POLYA, States::POLYA)       = 0.7;
-        //        tran_(States::POLYA, States::NONPOLYA)    = 1.0 - tran_(States::POLYA, States::POLYA);
-        //        tran_(States::NONPOLYA, States::POLYA)    = 0.0;
-        //        tran_(States::NONPOLYA, States::NONPOLYA) = 1.0 - tran_(States::NONPOLYA, States::POLYA);
-        return ret;
-    }
-
-    virtual bool write(const std::string& filename)
-    {
-        //		value_type t1 = tran_(States::POLYA, States::POLYA);
-        //		value_type t2 = tran_(States::NONPOLYA, States::POLYA);
-        // any chance of overwriting
-        //		tran_(States::POLYA, States::POLYA)       = 0.7;
-        //		tran_(States::POLYA, States::NONPOLYA)    = 1.0 - tran_(States::POLYA, States::POLYA);
-        //		tran_(States::NONPOLYA, States::POLYA)    = 0.0;
-        //		tran_(States::NONPOLYA, States::NONPOLYA) = 1.0 - tran_(States::NONPOLYA, States::POLYA);
-        bool ret = _base::write(filename);
-        // restore
-        //		tran_(States::POLYA, States::POLYA)       = t1;
-        //		tran_(States::POLYA, States::NONPOLYA)    = 1.0 - tran_(States::POLYA, States::POLYA);
-        //		tran_(States::NONPOLYA, States::POLYA)    = t2;
-        //		tran_(States::NONPOLYA, States::NONPOLYA) = 1.0 - tran_(States::NONPOLYA, States::POLYA);
-        return ret;
-    }
-
+    virtual bool read(const std::string& filename);
+    
+    virtual bool write(const std::string& filename);
+    
 /* evaluating algorithms */
 public:
     template <class TSequence>
@@ -190,30 +138,30 @@ auto PolyAHmmMode::calculateVirtabi(TIterator striter, size_t N) const -> const 
     matrix_type prob(no_states_, N);
     prob = 0.0;
     double curmax, tmp;
-    for (int i = 0; i < no_states_; ++i) {
-        prob(i, 0) = std::log2(init_(i, 0) * emit_(i, to_idx[*striter]));
+    for (size_t i = 0; i < no_states_; ++i) {
+        prob(i, 0) = std::log2(init_(i, 0) * emit_(i, to_idx[size_t(*striter)]));
     }
     // dynamically fill d
     ++striter; // at seq[1]
-    for (int j = 1; j < N; ++j, ++striter) {
+    for (size_t j = 1; j < N; ++j, ++striter) {
         // j is the observe sequence index
-        for (int i = 0; i < no_states_; ++i) {
+        for (size_t i = 0; i < no_states_; ++i) {
             // current state index i
             curmax = -INFINITY;
-            for (int k = 0; k < no_states_; ++k) { // previous state index k
+            for (size_t k = 0; k < no_states_; ++k) { // previous state index k
                 tmp = prob(k, j - 1) + std::log2(tran_(k, i));
                 if (tmp > curmax) {
                     curmax = tmp;
                 }
             }
-            prob(i, j) = curmax + std::log2(emit_(i, to_idx[*striter]));
+            prob(i, j) = curmax + std::log2(emit_(i, to_idx[size_t(*striter)]));
         }
     }
     path_.reSize(1, N);
     path_ = States::UNKNOWN;
     // determine the max ending
     curmax = -INFINITY;
-    for (int i = 0; i < no_states_; ++i) {
+    for (size_t i = 0; i < no_states_; ++i) {
         if (prob(i, N - 1) > curmax) {
             curmax = prob(i, N - 1);
             path_[N - 1] = i;
@@ -221,7 +169,7 @@ auto PolyAHmmMode::calculateVirtabi(TIterator striter, size_t N) const -> const 
     }
     for (int j = int(N - 2); j >= 0; --j) {
         curmax = -INFINITY;
-        for (int i = 0; i < no_states_; ++i) {
+        for (size_t i = 0; i < no_states_; ++i) {
             // from pos j (state: i) to pos j + 1 (state: path_(0, j+1) )
             tmp = prob(i, j) + std::log2(tran_(i, path_[j + 1]));
             if (tmp > curmax) {
@@ -252,22 +200,22 @@ auto PolyAHmmMode::calculateForward(TIterator striter, size_t N) const -> const 
     forw_.reSize(no_states_, N);
     forw_ = 0.0;
     // fill first sequence
-    for (int i = 0; i < no_states_; ++i) {
-        forw_(i, 0) = std::log2(init_(i, 0) * emit_(i, to_idx[*striter]));
+    for (size_t i = 0; i < no_states_; ++i) {
+        forw_(i, 0) = std::log2(init_(i, 0) * emit_(i, to_idx[size_t(*striter)]));
     }
     // dynamically fill d
     ++striter; // at seq[1]
     value_type logsum, temp;
-    for (int j = 1; j < N; ++j, ++striter) { // j is the observe sequence index
-        for (int i = 0; i < no_states_; ++i) { // current state index i
+    for (size_t j = 1; j < N; ++j, ++striter) { // j is the observe sequence index
+        for (size_t i = 0; i < no_states_; ++i) { // current state index i
             logsum = -INFINITY;
-            for (int k = 0; k < no_states_; ++k) { // previous state index
+            for (size_t k = 0; k < no_states_; ++k) { // previous state index
                 temp = forw_(k, j - 1) + std::log2(tran_(k, i));
                 if (temp > -INFINITY) {
                     logsum = temp + std::log2(1 + std::exp2(logsum - temp));
                 }
             }
-            forw_(i, j) = std::log2(emit_(i, to_idx[*striter])) + logsum;
+            forw_(i, j) = std::log2(emit_(i, to_idx[size_t(*striter)])) + logsum;
         }
     }
     return forw_;
@@ -291,18 +239,18 @@ auto PolyAHmmMode::calculateBackward(TIterator strriter, size_t N) const -> cons
     back_.reSize(no_states_, N);
     back_ = 0.0;
     // fill first sequence
-    for (int i = 0; i < no_states_; ++i) {
+    for (size_t i = 0; i < no_states_; ++i) {
         back_(i, N - 1) = 0.0 /* std::log2(1.) */;
     }
     value_type logsum, temp;
     for (int j = int(N - 2); j >= 0; --j, --strriter) {
         // state at j
-        for (int i = 0; i < no_states_; ++i) {
+        for (size_t i = 0; i < no_states_; ++i) {
             // i is on position j
             logsum = -INFINITY;
-            for (int k = 0; k < no_states_; ++k) {
+            for (size_t k = 0; k < no_states_; ++k) {
                 // k is on position j + 1
-                temp = back_(k, j + 1) + std::log2(tran_(i, k) * emit_(k, to_idx[*strriter]));
+                temp = back_(k, j + 1) + std::log2(tran_(i, k) * emit_(k, to_idx[size_t(*strriter)]));
                 if (temp > -INFINITY) {
                     logsum = temp + std::log2(1 + std::exp2(logsum - temp));
                 }
@@ -393,7 +341,7 @@ std::pair<size_t, size_t> PolyAHmmMode::maximumLikelihoodEstimationAux_(TSeqIter
         // *b: a Fasta;
         // b->seq_: a sequence
         for (auto ntiter = b->seq_.cbegin(); ntiter != b->seq_.cend(); ++ntiter) {
-            ++new_emit[to_idx[*ntiter]];
+            ++new_emit[to_idx[size_t(*ntiter)]];
             ++ret.second; // counting the # of nt
         }
         ++ret.first; // count the # of entries
