+++ src/quality.hpp
