+++ src/sequence.hpp
