+++ tests/data/polyA.fa
