+++ tests/data/polyA.fq
