+++ tests/data/polyA_train.fa
