+++ tests/files.cmake
@@ -1,0 +1,13 @@
+# Test case headers
+set(TrimIsoseqPolyA_Test_H
+    ${TrimIsoseqPolyA_TestsDir}/src/TestData.h
+)
+
+set(TrimIsoseqPolyA_Test_CPP
+    ${TrimIsoseqPolyA_TestsDir}/src/fasta_test.cpp
+    ${TrimIsoseqPolyA_TestsDir}/src/fastq_test.cpp
+    ${TrimIsoseqPolyA_TestsDir}/src/hmm_model_test.cpp
+    ${TrimIsoseqPolyA_TestsDir}/src/matrix_test.cpp
+    ${TrimIsoseqPolyA_TestsDir}/src/polyA_HMM_test.cpp
+    ${TrimIsoseqPolyA_TestsDir}/src/sequence_test.cpp
+)
